server/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Limiter==2.4.0
pymongo==3.12.0
PyJWT==2.3.0
gunicorn==20.1.0